use gfx.ts
